# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertKind": {
        "description": "Reasons the portfolio monitor raises an alert.\n\nEach is a *pre-committed rule the user wrote themselves* firing - not a\nrecommendation generated by the system.",
        "enum": [
          "profit_target_reached",
          "stop_loss_breached",
          "thesis_review_due",
          "position_concentration",
          "sector_concentration",
          "watchlist_entry_price_reached",
          "fundamental_red_flag"
        ],
        "title": "AlertKind",
        "type": "string"
      },
      "AlertRead": {
        "additionalProperties": false,
        "description": "A single alert.",
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_acknowledged": {
            "default": false,
            "title": "Is Acknowledged",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/AlertKind"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/AlertSeverity"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "id",
          "kind",
          "severity",
          "message",
          "created_at"
        ],
        "title": "AlertRead",
        "type": "object"
      },
      "AlertSeverity": {
        "description": "How loudly an alert should be surfaced in the UI.",
        "enum": [
          "info",
          "warning",
          "critical"
        ],
        "title": "AlertSeverity",
        "type": "string"
      },
...
      "CompanyDetail": {
        "additionalProperties": false,
        "description": "Single-company view, including the raw statements.",
        "properties": {
          "business_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Summary"
          },
          "financials": {
            "items": {
              "$ref": "#/components/schemas/AnnualFinancialsRead"
            },
            "title": "Financials",
            "type": "array"
          },
          "fiscal_years": {
            "items": {
              "type": "integer"
            },
            "title": "Fiscal Years",
            "type": "array"
          },
          "has_financials": {
            "default": false,
            "title": "Has Financials",
            "type": "boolean"
          },
          "has_price_history": {
            "default": false,
            "title": "Has Price History",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_close": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Most recent close held locally, if any.",
            "title": "Last Close"
          },
          "last_close_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Close Date"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sector": {
            "$ref": "#/components/schemas/Sector"
          },
          "sector_label": {
            "title": "Sector Label",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "required": [
          "id",
          "symbol",
          "name",
          "sector",
          "sector_label",
          "is_active"
        ],
        "title": "CompanyDetail",
        "type": "object"
      },
...
      "FundamentalsReport": {
        "additionalProperties": false,
        "description": "Full output of the fundamentals checklist for one company.",
        "properties": {
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "disclaimer": {
            "$ref": "#/components/schemas/Disclaimer",
            "default": {
              "is_financial_advice": false,
              "text": "Educational and general in nature. These figures score a company against a fixed checklist and do not predict price movements. Nothing here is financial advice, and no checklist guarantees profit or protects against loss. Consider a licensed financial advisor for your own situation."
            }
          },
          "fiscal_years": {
            "items": {
              "type": "integer"
            },
            "title": "Fiscal Years",
            "type": "array"
          },
          "latest_fiscal_year": {
            "title": "Latest Fiscal Year",
            "type": "integer"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricAssessment"
            },
            "title": "Metrics",
            "type": "array"
          },
          "peer_count": {
            "default": 0,
            "description": "Sector peers with usable data, used for median comparisons.",
            "title": "Peer Count",
            "type": "integer"
          },
          "red_flags": {
            "items": {
              "$ref": "#/components/schemas/RedFlag"
            },
            "title": "Red Flags",
            "type": "array"
          },
          "reference_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Price"
          },
          "reference_price_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Price Date"
          },
          "score": {
            "$ref": "#/components/schemas/FundamentalsScore"
          },
          "sector": {
            "$ref": "#/components/schemas/Sector"
          },
          "sector_label": {
            "title": "Sector Label",
            "type": "string"
          },
          "statement_review": {
            "$ref": "#/components/schemas/StatementReview"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "company_name",
          "sector",
          "sector_label",
          "fiscal_years",
          "latest_fiscal_year",
          "metrics",
          "statement_review",
          "red_flags",
          "score"
        ],
        "title": "FundamentalsReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_CompanySummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompanySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total number of matching rows, ignoring pagination.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CompanySummary]",
        "type": "object"
      },
...
      "PriceHistoryRead": {
        "additionalProperties": false,
        "description": "A company's stored price series.",
        "properties": {
          "bars": {
            "items": {
              "$ref": "#/components/schemas/PriceBarRead"
            },
            "title": "Bars",
            "type": "array"
          },
          "sessions": {
            "title": "Sessions",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "sessions",
          "bars"
        ],
        "title": "PriceHistoryRead",
        "type": "object"
      },
...
      "RedFlag": {
        "additionalProperties": false,
        "description": "A specific, evidenced concern - not a generic warning.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "severity": {
            "description": "'info', 'warning' or 'critical'.",
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title",
          "detail",
          "severity"
        ],
        "title": "RedFlag",
        "type": "object"
      },
...
      "Sector": {
        "description": "PSX sectors used for peer comparison and diversification checks.\n\nDeliberately coarse: the guide's advice is to compare P/E and debt against\npeers \"in the same sector\" and to diversify across sectors such as banks,\ncement, oil & gas, fertilizer and consumer goods.",
        "enum": [
          "commercial_banks",
          "cement",
          "oil_and_gas",
          "fertilizer",
          "food_and_personal_care",
          "power_generation",
          "textile_composite",
          "automobile_assembler",
          "technology_and_communication",
          "pharmaceuticals",
          "chemical",
          "other"
        ],
        "title": "Sector",
        "type": "string"
      },
...
      "TechnicalReport": {
        "additionalProperties": false,
        "description": "Timing context for a company - explicitly not a reason to buy or sell.\n\nThe guide is unambiguous that indicators are inputs, not verdicts, and that a\nbearish reading on a fundamentally strong stock held for years matters far\nless than the same reading on a short-term trade. :attr:`horizon_note`\ncarries that framing into the payload, personalised to the user's declared\ntime horizon when one is available.",
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "disclaimer": {
            "$ref": "#/components/schemas/Disclaimer",
            "default": {
              "is_financial_advice": false,
              "text": "Educational and general in nature. These figures score a company against a fixed checklist and do not predict price movements. Nothing here is financial advice, and no checklist guarantees profit or protects against loss. Consider a licensed financial advisor for your own situation."
            }
          },
          "horizon_note": {
            "default": "Technical readings help with timing, not with what to own. Treat them as context alongside the fundamentals, never as a signal on their own.",
            "title": "Horizon Note",
            "type": "string"
          },
          "last_close": {
            "title": "Last Close",
            "type": "number"
          },
          "moving_average_position": {
            "$ref": "#/components/schemas/MovingAveragePosition"
          },
          "moving_averages": {
            "$ref": "#/components/schemas/IndicatorReading"
          },
          "rsi": {
            "$ref": "#/components/schemas/IndicatorReading"
          },
          "rsi_zone": {
            "$ref": "#/components/schemas/RsiZone"
          },
          "sessions_analysed": {
            "title": "Sessions Analysed",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "trend": {
            "$ref": "#/components/schemas/IndicatorReading"
          },
          "trend_direction": {
            "$ref": "#/components/schemas/TrendDirection"
          },
          "volume": {
            "$ref": "#/components/schemas/IndicatorReading"
          },
          "volume_confirmation": {
            "$ref": "#/components/schemas/VolumeConfirmation"
          }
        },
        "required": [
          "symbol",
          "as_of",
          "sessions_analysed",
          "last_close",
          "trend",
          "rsi",
          "moving_averages",
          "volume",
          "trend_direction",
          "rsi_zone",
          "moving_average_position",
          "volume_confirmation"
        ],
        "title": "TechnicalReport",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDecision-support API for evaluating Pakistan Stock Exchange (PSX) equities.\n\n**What this API does:** scores a company against a fixed, published fundamentals\nchecklist; reports technical readings as timing context; records the pre-buy\nchecklist, profit target and stop-loss you commit to *before* buying; tracks your\nportfolio against the diversification limits you set for yourself; and tells you\nwhen one of your own rules has been crossed.\n\n**What it does not do:** predict prices, rate stocks, or tell you what to buy or\nsell. There is no endpoint that returns a recommendation, and that is a deliberate\nproduct constraint rather than a missing feature.\n\nCheck `GET /api/v1/meta` for `provider.is_synthetic`. When it is `true`, every\nfinancial figure and price served by this API is **generated demonstration data**\nfor real PSX ticker symbols - useful for exercising the application, useless for\nreal decisions.\n",
    "title": "PSX Invest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "description": "Stored alerts, newest first.\n\nRead-only: it reports what the last evaluation found. Call\n``POST /alerts/evaluate`` to re-check the portfolio against your rules.",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "description": "Include alerts you have already dismissed.",
            "in": "query",
            "name": "include_acknowledged",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include alerts you have already dismissed.",
              "title": "Include Acknowledged",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRead"
                  },
                  "title": "Response List Alerts Api V1 Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List alerts",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/api/v1/alerts/{alert_id}/acknowledge": {
      "post": {
        "description": "Dismiss an alert. The row is kept as part of your decision journal.",
        "operationId": "acknowledge_alert_api_v1_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "description": "Alert id.",
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "description": "Alert id.",
              "minimum": 1,
              "title": "Alert Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Alert not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss one alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/v1/companies": {
      "get": {
        "description": "Browse the companies available for analysis.",
        "operationId": "list_companies_api_v1_companies_get",
        "parameters": [
          {
            "description": "Restrict results to one PSX sector.",
            "in": "query",
            "name": "sector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Sector"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict results to one PSX sector.",
              "title": "Sector"
            }
          },
          {
            "description": "Match against symbol or company name.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Match against symbol or company name.",
              "title": "Search"
            }
          },
          {
            "description": "Rows per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Rows per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CompanySummary_"
                }
              }
            },
            "description": "A page of companies with their latest stored close."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List companies",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{symbol}": {
      "get": {
        "description": "One company with its stored annual figures, exactly as reported.\n\nDerived ratios are deliberately not included here - see the fundamentals\nendpoint, which explains and judges them.",
        "operationId": "get_company_api_v1_companies__symbol__get",
        "parameters": [
          {
            "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
              "examples": [
                "OGDC"
              ],
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown symbol."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get one company",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{symbol}/fundamentals": {
      "get": {
        "description": "Score a company against the guide's fundamentals checklist.\n\nReturns a verdict per metric, the four-question statement review, any red\nflags, and counts of how many criteria are met. It does not return a rating,\na price target or a buy/sell signal, and it never will - see\n``docs/ARCHITECTURE.md`` on why that constraint is treated as a requirement.",
        "operationId": "get_fundamentals_api_v1_companies__symbol__fundamentals_get",
        "parameters": [
          {
            "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
              "examples": [
                "OGDC"
              ],
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundamentalsReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown symbol."
          },
          "422": {
            "description": "No financial statements are stored for this company."
          }
        },
        "summary": "Run the fundamentals checklist",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{symbol}/prices": {
      "get": {
        "description": "Daily OHLCV bars, oldest first.",
        "operationId": "get_price_history_api_v1_companies__symbol__prices_get",
        "parameters": [
          {
            "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
              "examples": [
                "OGDC"
              ],
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "description": "Number of most recent sessions to return.",
            "in": "query",
            "name": "sessions",
            "required": false,
            "schema": {
              "default": 260,
              "description": "Number of most recent sessions to return.",
              "maximum": 1000,
              "minimum": 2,
              "title": "Sessions",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceHistoryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get price history",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{symbol}/technicals": {
      "get": {
        "description": "Trend, RSI, moving averages and volume confirmation for one company.\n\nThe user's declared time horizon is read from their profile purely to frame\nthe result: the guide's point is that a bearish reading matters far less to a\nfive-year holder than to a short-term trader. The numbers are identical either\nway.",
        "operationId": "get_technicals_api_v1_companies__symbol__technicals_get",
        "parameters": [
          {
            "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "PSX ticker symbol, e.g. OGDC. Case-insensitive.",
              "examples": [
                "OGDC"
              ],
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown symbol."
          },
          "422": {
            "description": "Not enough price history for a technical reading."
          }
        },
        "summary": "Get technical readings",
        "tags": [
          "companies"
        ]
      }
    },
...
  }
}